// src/app/honor-certificate/models/global-certificate.ts
export interface GlobalCertificate {
    step1: {
      lastName: string;
      firstName: string;
    };
    step2: {
      lastName: string;
      firstName: string;
      birthDate: string;
      customerNumber: string;
    };
    step3: {
      lastName: string;
      usageName: string;
      firstName: string;
      birthDate: string;
      birthCity: string;
      birthCountry: string;
      address: string;
      addressDetails: string;
      postalCode: string;
      city: string;
      country: string;
    };
    step4: {
      option: string;
    };
  }
  